requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# 数据库
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
influxdb-client>=1.43.0
qdrant-client>=1.9.0
redis[hiredis]>=5.0.0

# AI / LLM
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
tiktoken>=0.7.0

# 本地 Embedding + Reranker（USE_LOCAL_EMBEDDING=true 或 USE_RERANKER=true 时需要）
# sentence-transformers>=2.7.0  # 取消注释以启用本地 bge-m3 + CrossEncoder

# OCR
paddlepaddle>=2.6.0; sys_platform != "darwin"
paddleocr>=2.7.3

# 任务队列
celery[redis]>=5.4.0
watchfiles>=0.21.0        # dev 热更新 worker

# 工具
python-multipart>=0.0.9   # 文件上传
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
Pillow>=10.3.0
python-dotenv>=1.0.0
structlog>=24.2.0
